First cut at saving/restoring MIPS floating-point registers.
For newabi this is fairly simple as we just save/restore the 32
floating-point registers as doubles.  For O32 MIPS provides a variety
of floating-point ABIs.  For O32 MIPS with 64-bit floating-point
registers, save/restore the 32 floating-point registers as doubles.
For O32 MIPS with 32-bit floating-point registers, save/restore the 16
even registers as doubles.  This probably isn't perfect but there
isn't a predefined processor macro to determine which O32
floating-point ABI is in use aside from the register size.